+++ src/test/scala/com/vishnuvyas/classifiers/svm/DiabetesDataset.scala
@@ -10,7 +10,7 @@ import com.vishnuvyas.classifiers.Classifier
 /**
  * Created by vishnu on 4/20/14.
  */
-class DiabetesDataset extends AbstractDataset[Array[Double],java.lang.Boolean] {
+class DiabetesDataset extends AbstractDataset[Array[Float],java.lang.Boolean] {
 
   val dataFileStream = getClass.getResourceAsStream("/pima-indians-diabetes.data")
   val fieldNames = Array("npreg","glucose","dbp","tricep","insulin2hr","bmi","dbpf","age","diabetec")
@@ -30,15 +30,15 @@ class DiabetesDataset extends AbstractDataset[Array[Double],java.lang.Boolean] {
 
   def initialize() = {
     val rawdataset = Source.fromInputStream(dataFileStream).getLines().map { line =>
-      val a = line.trim().split(",").map(_.toDouble)
+      val a = line.trim().split(",").map(_.toFloat)
       val q = a.take(8).flatMap(ai => a.take(8).map(aj => ai*aj))
       Tuple2(a.take(8),(if (a(8) == 1) java.lang.Boolean.TRUE else java.lang.Boolean.FALSE))
     }.toList
 
     // now center the dataset by computing component wise means
-    val scaleInfo = rawdataset.foldLeft(Map[Int,(Double,Double,Int)]()) { (m,d) =>
+    val scaleInfo = rawdataset.foldLeft(Map[Int,(Float,Float,Int)]()) { (m,d) =>
       d._1.zipWithIndex.foldLeft(m) { (mi,f) =>
-        val default = mi.getOrElse(f._2,(0.0,0.0,0))
+        val default = mi.getOrElse(f._2,(0.0f,0.0f,0))
         mi + (f._2 -> (f._1+default._1,f._1*f._1 + default._2, 1 + default._3))
       }
     }
@@ -49,7 +49,7 @@ class DiabetesDataset extends AbstractDataset[Array[Double],java.lang.Boolean] {
 
     val rescaledDataset = rawdataset.map { d =>
       val rescaledVector = d._1.zipWithIndex.map { case (v,f) =>
-        (v - rescaleInfo(f)._1)/rescaleInfo(f)._2
+        ((v - rescaleInfo(f)._1)/rescaleInfo(f)._2).toFloat
       }
       (rescaledVector -> d._2)
     }
@@ -57,7 +57,7 @@ class DiabetesDataset extends AbstractDataset[Array[Double],java.lang.Boolean] {
     rescaledDataset
   }
 
-  override def getPoint(i: Int): Array[Double] = dataset(i)._1
+  override def getPoint(i: Int): Array[Float] = dataset(i)._1
 
   override def getLabel(i: Int): java.lang.Boolean = dataset(i)._2
 
@@ -67,20 +67,20 @@ class DiabetesDataset extends AbstractDataset[Array[Double],java.lang.Boolean] {
 
   override def count(): Int = dataset.size
 
-  override def sample(fraction: Double): Dataset[Array[Double], java.lang.Boolean] = new Dataset[Array[Double],java.lang.Boolean] {
-    val sampleDataset = dataset.filter(_ => Random.nextDouble() < fraction)
+  override def sample(fraction: Float): Dataset[Array[Float], java.lang.Boolean] = new Dataset[Array[Float],java.lang.Boolean] {
+    val sampleDataset = dataset.filter(_ => Random.nextFloat() < fraction)
 
     override def count(): Int = sampleDataset.size
-    override def getPoint(i: Int) : Array[Double] = sampleDataset(i)._1
+    override def getPoint(i: Int) : Array[Float] = sampleDataset(i)._1
     override def getLabel(i : Int) : java.lang.Boolean = sampleDataset(i)._2
     override def getLabels : util.Set[java.lang.Boolean] = Set(java.lang.Boolean.TRUE,java.lang.Boolean.FALSE)
     override def dim(): Int = 8
 
-    override def sample(fraction: Double): Dataset[Array[Double], java.lang.Boolean] = ???
+    override def sample(fraction: Float): Dataset[Array[Float], java.lang.Boolean] = ???
   }
 
 
-  override def trainingError(c:Classifier[Array[Double],java.lang.Boolean]):Int = {
+  override def trainingError(c:Classifier[Array[Float],java.lang.Boolean]):Int = {
     dataset.map(_ match {
       case (xi,yi) =>
         val (prediction,margin) = c.predictions(xi).toSeq.sortWith(_._2 < _._2)(0)
@@ -88,7 +88,7 @@ class DiabetesDataset extends AbstractDataset[Array[Double],java.lang.Boolean] {
     }).sum
   }
 
-  def confusionMatrix(c:Classifier[Array[Double],java.lang.Boolean]) = {
+  def confusionMatrix(c:Classifier[Array[Float],java.lang.Boolean]) = {
     var (tp,fp,tn,fn) = (0,0,0,0)
     dataset.foreach {
       case (xi,yi) =>
